split related components
